Wood StockPile

+++ wood_stockpile/lib/intermediate_screen.dart
@@ -14,6 +14,154 @@ var status = false;
 var email = "";
 var type = "";
 
+// class IntermediateScreen extends StatefulWidget {
+//   @override
+//   State<IntermediateScreen> createState() => _IntermediateScreenState();
+// }
+
+// class _IntermediateScreenState extends State<IntermediateScreen> {
+//   @override
+//   void initState() {
+//     super.initState();
+//     fetchData();
+//   }
+
+//   Future<void> fetchData() async {
+//     final SharedPreferences prefs = await SharedPreferences.getInstance();
+//     try {
+//       print(_firebase.currentUser!.uid);
+//       final DocumentSnapshot snapshot1 = await FirebaseFirestore.instance
+//           .collection("unverified_users")
+//           .doc(_firebase.currentUser!.uid)
+//           .get();
+
+//       if (snapshot1.exists) {
+//         final Map<String, dynamic> data =
+//             snapshot1.data() as Map<String, dynamic>;
+
+//         // Set data in 'users' collection
+//         await FirebaseFirestore.instance
+//             .collection('users')
+//             .doc(_firebase.currentUser!.uid)
+//             .set({
+//           'email': data['email'],
+//           'password': data['password'],
+//           'mobile_number': data['mobile_number'],
+//           'type': data['type'],
+//           'username': data['username'],
+//         });
+
+//         // Delete document from 'unverified_users'
+//         await FirebaseFirestore.instance
+//             .collection('unverified_users')
+//             .doc(_firebase.currentUser!.uid)
+//             .delete();
+//         prefs.setString("unumber", data['mobile_number']);
+//         prefs.setString("uemail", data['email']);
+//         print("Data moved and deleted successfully.");
+//       } else {
+//         print("Document not found in 'unverified_users'.");
+//       }
+// //--------------------------------------------------------------
+//       final DocumentSnapshot snapshot = await FirebaseFirestore.instance
+//           .collection("users")
+//           .doc(_firebase.currentUser!.uid)
+//           .get();
+
+//       if (snapshot.exists) {
+//         final Map<String, dynamic> data =
+//             snapshot.data() as Map<String, dynamic>;
+//         prefs.setString("unumber", data['mobile_number'] as String);
+//         prefs.setString("uemail", data['email'] as String);
+
+//         final String? fetchedEmail = data['email'] as String?;
+//         final String? fetchedtype = data['type'] as String?;
+//         // if (fetchedEmail != null && fetchedEmail.isNotEmpty) {
+//         //   setState(() {
+//         //     status = true;
+//         //     email = fetchedEmail;
+//         //     type = fetchedtype.toString();
+//         //   });
+//         //   if (type == "user") {
+//         //     Navigator.pushReplacement(
+//         //         context, MaterialPageRoute(builder: (context) => Homescreen()));
+//         //   }
+//         //   if (type == "wood_seller") {
+//         //     await prefs.setString('company_email', fetchedEmail);
+//         //     Navigator.pushReplacement(
+//         //         context, MaterialPageRoute(builder: (context) => homescreen()));
+//         //   }
+//         // }
+//         if (fetchedEmail != null && fetchedEmail.isNotEmpty) {
+//           setState(() {
+//             status = true;
+//             email = fetchedEmail;
+//             type = data['type'].toString(); // Safely access 'type'
+//           });
+
+//           if (type == "user") {
+//             Navigator.pushReplacement(
+//                 context, MaterialPageRoute(builder: (context) => Homescreen()));
+//           }
+//           if (type == "wood_seller") {
+//             await prefs.setString('company_email', fetchedEmail);
+//             Navigator.pushReplacement(
+//                 context, MaterialPageRoute(builder: (context) => homescreen()));
+//           }
+//         } else {
+//           // Handle the case where 'email' is empty or null
+//           // Navigator.pushReplacement(
+//           //     context, MaterialPageRoute(builder: (context) => login()));
+//         }
+//       } else {
+//         // Handle the case where the document doesn't exist
+//         final DocumentSnapshot snapshot2 = await FirebaseFirestore.instance
+//             .collection("unverified_wood_seller")
+//             .doc(_firebase.currentUser!.uid)
+//             .get();
+//         final Map<String, dynamic> data =
+//             snapshot2.data() as Map<String, dynamic>;
+//         final String? fetchedEmail = data['email'] as String?;
+//         print(fetchedEmail);
+//         if (fetchedEmail != null && fetchedEmail.isNotEmpty) {
+//           Navigator.pushReplacement(context,
+//               MaterialPageRoute(builder: (context) => waiting_screen()));
+//         }
+//       }
+//     } catch (error) {
+//       // Handle errors that may occur during data retrieval
+//       print("Error: $error");
+//     }
+//   }
+
+//   void backtologin() {
+//     Navigator.push(context, MaterialPageRoute(builder: (context) => login()));
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       body: Column(
+//         mainAxisAlignment: MainAxisAlignment.center,
+//         children: [
+//           Center(
+//               child: SizedBox(
+//             height: 400, // Adjust the height as needed
+//             child: Lottie.asset("assets/lottie/loading.json"),
+//           )),
+//           // ElevatedButton.icon(
+//           //     style: ElevatedButton.styleFrom(
+//           //       backgroundColor: Colors.black, // background color
+//           //       foregroundColor: Colors.white,
+//           //     ),
+//           //     onPressed: backtologin,
+//           //     icon: Icon(Icons.arrow_back),
+//           //     label: Text("Back to Login Screen"))
+//         ],
+//       ),
+//     );
+//   }
+// }
 class IntermediateScreen extends StatefulWidget {
   @override
   State<IntermediateScreen> createState() => _IntermediateScreenState();
@@ -27,62 +175,71 @@ class _IntermediateScreenState extends State<IntermediateScreen> {
   }
 
   Future<void> fetchData() async {
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
     try {
-      print(_firebase.currentUser!.uid);
       final DocumentSnapshot snapshot1 = await FirebaseFirestore.instance
           .collection("unverified_users")
           .doc(_firebase.currentUser!.uid)
           .get();
 
       if (snapshot1.exists) {
-        final Map<String, dynamic> data =
-            snapshot1.data() as Map<String, dynamic>;
+        final Map<String, dynamic>? data =
+            snapshot1.data() as Map<String, dynamic>?;
+        if (data != null) {
+          await FirebaseFirestore.instance
+              .collection('users')
+              .doc(_firebase.currentUser!.uid)
+              .set({
+            'email': data['email'],
+            'password': data['password'],
+            'mobile_number': data['mobile_number'],
+            'type': data['type'],
+            'username': data['username'],
+          });
 
-        // Set data in 'users' collection
-        await FirebaseFirestore.instance
-            .collection('users')
-            .doc(_firebase.currentUser!.uid)
-            .set({
-          'email': data['email'],
-          'password': data['password'],
-          'mobile_number': data['mobile_number'],
-          'type': data['type'],
-          'username': data['username'],
-        });
+          await FirebaseFirestore.instance
+              .collection('unverified_users')
+              .doc(_firebase.currentUser!.uid)
+              .delete();
+          prefs.setString("unumber", data['mobile_number'] ?? '');
+          prefs.setString("uemail", data['email'] ?? '');
 
-        // Delete document from 'unverified_users'
-        await FirebaseFirestore.instance
-            .collection('unverified_users')
-            .doc(_firebase.currentUser!.uid)
-            .delete();
-
-        print("Data moved and deleted successfully.");
+          prefs.setString("uname", data["username"] ?? '');
+          print("Data moved and deleted successfully.");
+        } else {
+          print("Document data in 'unverified_users' is null.");
+        }
       } else {
         print("Document not found in 'unverified_users'.");
       }
-//--------------------------------------------------------------
-      final SharedPreferences prefs = await SharedPreferences.getInstance();
+
       final DocumentSnapshot snapshot = await FirebaseFirestore.instance
           .collection("users")
           .doc(_firebase.currentUser!.uid)
           .get();
 
       if (snapshot.exists) {
-        final Map<String, dynamic> data =
-            snapshot.data() as Map<String, dynamic>;
-        final String? fetchedEmail = data['email'] as String?;
-        final String? fetchedtype = data['type'] as String?;
-        if (fetchedEmail != null && fetchedEmail.isNotEmpty) {
+        final Map<String, dynamic>? data =
+            snapshot.data() as Map<String, dynamic>?;
+        final String? fetchedEmail = data?['email'];
+        final String? fetchedType = data?['type'];
+
+        if (fetchedEmail != null &&
+            fetchedEmail.isNotEmpty &&
+            fetchedType != null) {
           setState(() {
             status = true;
             email = fetchedEmail;
-            type = fetchedtype.toString();
+            type = fetchedType.toString();
           });
+
           if (type == "user") {
+            prefs.setString("unumber", data?['mobile_number'] ?? '');
+            prefs.setString("uemail", data?['email'] ?? '');
+            prefs.setString("uname", data?["username"] ?? '');
             Navigator.pushReplacement(
                 context, MaterialPageRoute(builder: (context) => Homescreen()));
-          }
-          if (type == "wood_seller") {
+          } else if (type == "wood_seller") {
             await prefs.setString('company_email', fetchedEmail);
             Navigator.pushReplacement(
                 context, MaterialPageRoute(builder: (context) => homescreen()));
@@ -93,14 +250,13 @@ class _IntermediateScreenState extends State<IntermediateScreen> {
           //     context, MaterialPageRoute(builder: (context) => login()));
         }
       } else {
-        // Handle the case where the document doesn't exist
         final DocumentSnapshot snapshot2 = await FirebaseFirestore.instance
             .collection("unverified_wood_seller")
             .doc(_firebase.currentUser!.uid)
             .get();
-        final Map<String, dynamic> data =
-            snapshot2.data() as Map<String, dynamic>;
-        final String? fetchedEmail = data['email'] as String?;
+        final Map<String, dynamic>? data =
+            snapshot2.data() as Map<String, dynamic>?;
+        final String? fetchedEmail = data?['email'];
         print(fetchedEmail);
         if (fetchedEmail != null && fetchedEmail.isNotEmpty) {
           Navigator.pushReplacement(context,
@@ -108,12 +264,11 @@ class _IntermediateScreenState extends State<IntermediateScreen> {
         }
       }
     } catch (error) {
-      // Handle errors that may occur during data retrieval
       print("Error: $error");
     }
   }
 
-  void backtologin() {
+  void backToLogin() {
     Navigator.push(context, MaterialPageRoute(builder: (context) => login()));
   }
 
@@ -124,16 +279,17 @@ class _IntermediateScreenState extends State<IntermediateScreen> {
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
           Center(
-              child: SizedBox(
-            height: 400, // Adjust the height as needed
-            child: Lottie.asset("assets/lottie/loading.json"),
-          )),
+            child: SizedBox(
+              height: 400, // Adjust the height as needed
+              child: Lottie.asset("assets/lottie/loading.json"),
+            ),
+          ),
           // ElevatedButton.icon(
           //     style: ElevatedButton.styleFrom(
           //       backgroundColor: Colors.black, // background color
           //       foregroundColor: Colors.white,
           //     ),
-          //     onPressed: backtologin,
+          //     onPressed: backToLogin,
           //     icon: Icon(Icons.arrow_back),
           //     label: Text("Back to Login Screen"))
         ],
